refactor duplicated functions into standalone modules
