+++ plutarch-quickcheck.cabal
@@ -105,6 +105,10 @@ library
   import:          common-lang
   exposed-modules: Plutarch.Test.QuickCheck
   other-modules:
+    Plutarch.Extra.List
+    Plutarch.Extra.Map
+    Plutarch.Extra.Map.Unsorted
+    Plutarch.Extra.Maybe
     Plutarch.Test.QuickCheck.Function
     Plutarch.Test.QuickCheck.Instances
 
@@ -122,6 +126,7 @@ library
     , tagged
     , tasty
     , tasty-expected-failure
+    , tasty-hunit
     , tasty-quickcheck
     , text
     , universe
